Update common-issues.md
As of pass 1.7.4 I had to use a second key and re-init back to the original to force re-encryption. Finally figured this out after wondering why nothing changed (since the key ID didn't change, I didn't know if the key had changed -- apparently it did and kept the same ID, because reencrypting fixed the android app).